// vendor/project-e-protocol/generated/typescript/project_e/transfer/v1/content_pb.ts
// @generated by protoc-gen-es v2.12.1 with parameter "target=ts,import_extension=js"
// @generated from file project_e/transfer/v1/content.proto (package project_e.transfer.v1, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file project_e/transfer/v1/content.proto.
 */
export const file_project_e_transfer_v1_content: GenFile = /*@__PURE__*/
  fileDesc("CiNwcm9qZWN0X2UvdHJhbnNmZXIvdjEvY29udGVudC5wcm90bxIVcHJvamVjdF9lLnRyYW5zZmVyLnYxIp4BChFDb250ZW50RGVzY3JpcHRvchIRCglmaWxlX25hbWUYASABKAkSFwoKbWVkaWFfdHlwZRgCIAEoCUgAiAEBEgwKBHNpemUYAyABKAQSDgoGYmxha2UzGAQgASgMEjAKBGtpbmQYBSABKA4yIi5wcm9qZWN0X2UudHJhbnNmZXIudjEuQ29udGVudEtpbmRCDQoLX21lZGlhX3R5cGUqdgoLQ29udGVudEtpbmQSHAoYQ09OVEVOVF9LSU5EX1VOU1BFQ0lGSUVEEAASFQoRQ09OVEVOVF9LSU5EX0ZJTEUQARIaChZDT05URU5UX0tJTkRfVEVYVF9VVEY4EAISFgoSQ09OVEVOVF9LSU5EX0lNQUdFEANCBroCA1BKRWIGcHJvdG8z");

/**
 * @generated from message project_e.transfer.v1.ContentDescriptor
 */
export type ContentDescriptor = Message<"project_e.transfer.v1.ContentDescriptor"> & {
  /**
   * @generated from field: string file_name = 1;
   */
  fileName: string;

  /**
   * @generated from field: optional string media_type = 2;
   */
  mediaType?: string | undefined;

  /**
   * @generated from field: uint64 size = 3;
   */
  size: bigint;

  /**
   * @generated from field: bytes blake3 = 4;
   */
  blake3: Uint8Array;

  /**
   * @generated from field: project_e.transfer.v1.ContentKind kind = 5;
   */
  kind: ContentKind;
};

/**
 * Describes the message project_e.transfer.v1.ContentDescriptor.
 * Use `create(ContentDescriptorSchema)` to create a new message.
 */
export const ContentDescriptorSchema: GenMessage<ContentDescriptor> = /*@__PURE__*/
  messageDesc(file_project_e_transfer_v1_content, 0);

/**
 * @generated from enum project_e.transfer.v1.ContentKind
 */
export enum ContentKind {
  /**
   * @generated from enum value: CONTENT_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: CONTENT_KIND_FILE = 1;
   */
  FILE = 1,

  /**
   * @generated from enum value: CONTENT_KIND_TEXT_UTF8 = 2;
   */
  TEXT_UTF8 = 2,

  /**
   * @generated from enum value: CONTENT_KIND_IMAGE = 3;
   */
  IMAGE = 3,
}

/**
 * Describes the enum project_e.transfer.v1.ContentKind.
 */
export const ContentKindSchema: GenEnum<ContentKind> = /*@__PURE__*/
  enumDesc(file_project_e_transfer_v1_content, 0);